+++ src/main/java/dev/gacbl/logicore/client/ClientKnowledgeData.java
@@ -22,4 +22,8 @@ public class ClientKnowledgeData {
     public static void addAll(List<String> itemKeys) {
         UNLOCKED_ITEMS.addAll(itemKeys);
     }
+
+    public static Set<String> getUnlockedItems() {
+        return Set.copyOf(UNLOCKED_ITEMS);
+    }
 }